Add files via upload
This project involves creating an interactive interface where users respond to visual stimuli by clicking a button as quickly as possible .The user is presented with a moving box on the screen and must click in it as quickly as possible . The user is presented with a moving box on the screen and must click on it as quickly as possible. After each successful click, the application calculates and displays the user's reaction time, providing immediate feedback. This project enhances user engagement through a fun and challenging experience while offering insights into their reaction speed.